fix bf16 dtype mismatch in all interp runners, add SAE loss chart
Models like Qwen output bf16 activations but SAE/probe weights are float32.
Cast activations to float32 before passing to nn.Linear/SAE/SVD in all
runners (SAE train, SAE analyze, linear probe, PCA, steer apply hook).
Add training loss curve chart to SAE train results view.

+++ src/serve/sae_analyze_runner.py
@@ -40,8 +40,8 @@ def run_sae_analyze(options: SaeAnalyzeOptions) -> dict[str, Any]:
             model(ids)
             acts = extractor.get_activations()[target_layer]
 
-    # Mean-pool across sequence
-    pooled = acts[0].mean(dim=0).unsqueeze(0)
+    # Mean-pool across sequence, cast to SAE dtype (model may be bf16)
+    pooled = acts[0].mean(dim=0).unsqueeze(0).to(dtype=torch.float32)
 
     with torch.no_grad():
         reconstruction, latent = sae(pooled)
